Use common types from objc2::foundation
NSUInteger is now usize, so a bunch of `as u64` casts were removed

+++ examples/compute/compute-argument-buffer.rs
@@ -23,7 +23,7 @@ fn main() {
 
         let buffer = device.new_buffer_with_data(
             unsafe { mem::transmute(data.as_ptr()) },
-            (data.len() * mem::size_of::<u32>()) as u64,
+            data.len() * mem::size_of::<u32>(),
             MTLResourceOptions::CPUCacheModeDefaultCache,
         );
 
@@ -31,7 +31,7 @@ fn main() {
             let data = [0u32];
             device.new_buffer_with_data(
                 unsafe { mem::transmute(data.as_ptr()) },
-                (data.len() * mem::size_of::<u32>()) as u64,
+                data.len() * mem::size_of::<u32>(),
                 MTLResourceOptions::CPUCacheModeDefaultCache,
             )
         };
@@ -77,7 +77,7 @@ fn main() {
         };
 
         let thread_group_size = MTLSize {
-            width: (data.len() as u64 + width) / width,
+            width: (data.len() + width) / width,
             height: 1,
             depth: 1,
         };
